# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/audit-logs": {
      "get": {
        "operationId": "AdminComplianceController_getAuditLogs_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/audit-logs/{id}": {
      "get": {
        "operationId": "AdminComplianceController_getAuditLog_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/compliance/data-requests": {
      "get": {
        "operationId": "AdminComplianceController_listDataRequests_v1",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/compliance/data-requests/{id}/process": {
      "post": {
        "operationId": "AdminComplianceController_processDataRequest_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/compliance/retention": {
      "get": {
        "operationId": "AdminComplianceController_getRetentionPolicy_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/roles": {
      "get": {
        "operationId": "AdminComplianceController_getRoles_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "operationId": "AdminComplianceController_createRole_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/roles/{id}": {
      "patch": {
        "operationId": "AdminComplianceController_updateRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/roles": {
      "post": {
        "operationId": "AdminComplianceController_assignRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/tax/vat-records": {
      "get": {
        "operationId": "AdminComplianceController_getVatRecords_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/legal/pages": {
      "get": {
        "operationId": "AdminComplianceController_getLegalPages_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/legal/pages/{slug}": {
      "patch": {
        "operationId": "AdminComplianceController_updateLegalPage_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cadna Mart API",
    "description": "Backend API for the Cadna Mart marketplace.\n\n### Response envelope\nEvery successful response is wrapped by the global response interceptor:\n```json\n{\n  \"success\": true,\n  \"statusCode\": 200,\n  \"data\": { ... },\n  \"message\": \"...\",\n  \"meta\": { \"correlationId\": \"...\", \"timestamp\": \"...\" }\n}\n```\nWhen documenting individual endpoints below, the schema describes the `data` payload only.\n\n### Authentication\n- **Bearer JWT** (`Authorization: Bearer <accessToken>`) for authenticated user routes.\n- **Guest token** (`x-guest-token: <token>`) for guest-cart endpoints. Returned once from `POST /cart`.\n\n### Error shape\nErrors flow through the global `AllExceptionsFilter`:\n```json\n{\n  \"success\": false,\n  \"statusCode\": 400,\n  \"message\": \"Validation failed\",\n  \"errorCode\": \"BAD_REQUEST\",\n  \"path\": \"/api/v1/...\",\n  \"correlationId\": \"...\"\n}\n```",
    "version": "1.0",
    "contact": {
      "name": "Cadna Mart Engineering",
      "url": "",
      "email": "cadnamart@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "https://cadna-mart-be-nsz2.onrender.com/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned from /auth/login or /auth/otp/verify"
      },
      "guestToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-guest-token",
        "description": "Guest cart access token (returned once from POST /cart for guests)"
      }
    },
...
  }
}